Renamed the test class and the test data file, moved the test class to the folder field, added trim() and tolower() method
