+++ src/test/java/gov/cancer/tests/fields/SearchEngineRestrictions_Test.java
@@ -1,4 +1,4 @@
-package gov.cancer.tests.crosscutting;
+package gov.cancer.tests.fields;
 
 import java.util.Iterator;
 
@@ -7,12 +7,12 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import gov.cancer.framework.ExcelDataReader;
-import gov.cancer.pageobject.crosscutting.PageWithMetatag;
+import gov.cancer.pageobject.crosscutting.PageWithMetaTag;
 import gov.cancer.pageobject.helper.RobotMetaTag;
 import gov.cancer.tests.TestObjectBase;
 import gov.cancer.tests.TestRunner;
 
-public class PageWithMetatag_Test extends TestObjectBase {
+public class SearchEngineRestrictions_Test extends TestObjectBase {
 
   /**
    * Verify that the "robot" tag is present
@@ -23,7 +23,7 @@ public class PageWithMetatag_Test extends TestObjectBase {
   @Test(dataProvider = "getPagesWithIndex")
   public void verifyIncludeSearchIsSelected(String path) {
 
-    TestRunner.run(PageWithMetatag.class, path, (PageWithMetatag page) -> {
+    TestRunner.run(PageWithMetaTag.class, path, (PageWithMetaTag page) -> {
       RobotMetaTag rmt = page.getRobotMetaTag();
       Assert.assertTrue(rmt.getAllowSearch(), "Found index as Robot Metatag content");
 
@@ -41,7 +41,7 @@ public class PageWithMetatag_Test extends TestObjectBase {
   @Test(dataProvider = "getPagesWithNoIndex")
   public void verifyExcludeFromSearchIsSelected(String path) {
 
-    TestRunner.run(PageWithMetatag.class, path, (PageWithMetatag page) -> {
+    TestRunner.run(PageWithMetaTag.class, path, (PageWithMetaTag page) -> {
       RobotMetaTag rmt = page.getRobotMetaTag();
       Assert.assertFalse(rmt.getAllowSearch(), "Error: Found index as Robot Metatag content");
 
@@ -60,7 +60,7 @@ public class PageWithMetatag_Test extends TestObjectBase {
   @DataProvider(name = "getPagesWithIndex")
   public Iterator<Object[]> getPagesWithIndex() {
     String[] columns = { "path" };
-    return new ExcelDataReader(getDataFilePath("page-with-metatag-data.xlsx"), "getpages_with_index", columns);
+    return new ExcelDataReader(getDataFilePath("searchenginerestrictions_data.xlsx"), "getpages_with_index", columns);
   }
 
   /**
@@ -73,6 +73,6 @@ public class PageWithMetatag_Test extends TestObjectBase {
   @DataProvider(name = "getPagesWithNoIndex")
   public Iterator<Object[]> getPagesWithNoIndex() {
     String[] columns = { "path" };
-    return new ExcelDataReader(getDataFilePath("page-with-metatag-data.xlsx"), "getpages_with_noindex", columns);
+    return new ExcelDataReader(getDataFilePath("searchenginerestrictions_data.xlsx"), "getpages_with_noindex", columns);
   }
 }
